Update command-actions.md - Round #1 (#20)
* Update command-actions.md - Round #1

I didn't see the video portion in my files yet, so I did this edit without comparing directly to what's on the screen. Definitely double check the code/terminal portions and make sure they're correct - like anything that may need a "/", ".php", or "()", for example. I tweaked that in a couple of places, based on what I know about directories and methods, which may or may not have been necessary.

* small change

---------

Co-authored-by: Diego Aguiar <barret91@hotmail.com>