Update chain-of-responsibility.md - Round #1
Sorry it took me so long to get to this, guys! I had a massive headache Thursday/Friday and didn't get much done.

This sounded pretty good already. Just made some minor tweaks here and there to change up the flow. The code is usually spot on without my help. I just deleted a few semicolons since I'm used to omitting them in the script text, and I moved a few "/' to the left. That's where I usually put them in my edits, but I may not be correct.